+++ src/components/ModalCountryDetails/Gastronomy/Slider/Slider.tsx
