--- src/components/Slider/Slider.css
@@ -1,64 +1,0 @@
-.dishes {
-	display: flex;
-	flex-wrap: wrap;
-	width: 100%;
-	justify-content: space-around;
-	margin-top: 25px;
-	gap: 10px;
-}
-
-.dish-card-slider {
-	width: 250px;
-	height: 320px;
-	margin: auto;
-	border: 1px solid var(--main-color-modal);
-	border-radius: 20px;
-	box-shadow: 5px 0px 5px var(--main-color-modal);
-}
-
-article h1 {
-	width: 200px;
-	margin: auto;
-	background: linear-gradient(to bottom, rgb(66, 140, 149, 0.4), white);
-	font-family: var(--font-family-title-home);
-	color: var(--main-color-modal);
-	padding: 5px 5px 5px 5px;
-	border-radius: 10px;
-	box-shadow: 5px 5px 5px var(--main-color-modal);
-}
-
-article h2 {
-	font-family: var(--font-family-modal);
-	color: var(--color-text-modal);
-}
-
-article p {
-	font-family: var(--font-family-modal);
-	color: var(--color-text-modal);
-	width: 180px;
-	margin: auto;
-	padding-bottom: 20px;
-}
-
-.carousel #dish-undefined {
-	width: 180px;
-	height: 120px;
-}
-
-.carousel-status {
-	display: none;
-}
-
-.carousel .thumbs-wrapper {
-	display: none;
-}
-
-@media screen and (min-width: 768px) {
-	.carousel {
-		display: none;
-	}
-
-	article h1 {
-		width: 300px;
-	}
-}
